+++ embuilder.py
@@ -105,7 +105,7 @@ do that, run
    2. cmake . -DCMAKE_BUILD_TYPE=Release
    3. make (or mingw32-make/vcbuild/msbuild on Windows)
 
-and set up the location to the native optimizer in ~/.emscripten
+and set up the location to the native optimizer in .emscripten
 ''' % '\n        '.join(all_tasks)