+++ emcc.py
@@ -99,7 +99,8 @@ UNSUPPORTED_LLD_FLAGS = {
     '-soname': True,
     '-allow-shlib-undefined': False,
     '-rpath': True,
-    '-rpath-link': True
+    '-rpath-link': True,
+    '-version-script': True,
 }
 
 LIB_PREFIXES = ('', 'lib')
@@ -850,9 +851,9 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
             pass
 
     if run_via_emxx:
-      compiler = [shared.PYTHON, shared.EMXX]
+      compiler = [shared.EMXX]
     else:
-      compiler = [shared.PYTHON, shared.EMCC]
+      compiler = [shared.EMCC]
 
     cmd = compiler + [a for a in args if a != '--tracing']
     # configure tests want a more shell-like style, where we emit return codes on exit()
@@ -1258,6 +1259,11 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
 
     shared.verify_settings()
 
+    # We allow this warning to be supressed by the environment so that we can run the test
+    # suite against fastcomp while supressing this warning.
+    if not shared.Settings.WASM_BACKEND and 'EMCC_ALLOW_FASTCOMP' not in os.environ:
+      diagnostics.warning('fastcomp', 'the fastomp compiler is deprecated.  Please switch to the upstream llvm backend as soon as possible and open issues if you have trouble doing so')
+
     if options.no_entry or '_main' not in shared.Settings.EXPORTED_FUNCTIONS:
       shared.Settings.EXPECT_MAIN = 0
 
@@ -1408,7 +1414,9 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
     # We add these to the user's flags (newargs), but not when building .s or .S assembly files
     cflags = shared.get_cflags(newargs)
 
-    newargs = [x for x in newargs if x not in ['-msse']]
+    # SSEx is implemented on top of SIMD128 instruction set, but do not pass SSE flags to LLVM
+    # so it won't think about generating native x86 SSE code.
+    newargs = [x for x in newargs if x not in shared.SIMD_FEATURE_TOWER]
 
     if not shared.Settings.STRICT:
       # The preprocessor define EMSCRIPTEN is deprecated. Don't pass it to code
@@ -3066,7 +3074,7 @@ def parse_args(newargs):
         shared.generate_config(optarg)
       should_exit = True
     # Record SIMD setting for Binaryen
-    elif newargs[i] == '-msimd128':
+    elif newargs[i] == '-msimd128' or newargs[i] in shared.SIMD_FEATURE_TOWER:
       shared.Settings.BINARYEN_SIMD = 1
     # Record USE_PTHREADS setting because it controls whether --shared-memory is passed to lld
     elif newargs[i] == '-pthread':
@@ -3311,7 +3319,7 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
                                       opt_level=shared.Settings.OPT_LEVEL,
                                       minify_whitespace=optimizer.minify_whitespace,
                                       use_closure_compiler=options.use_closure_compiler,
-                                      debug_info=intermediate_debug_info,
+                                      debug_info=debug_info,
                                       symbols_file=symbols_file)
     if shared.Settings.WASM == 2:
       shutil.copyfile(wasm2js, wasm2js_template)