+++ emscripten.py
@@ -1582,7 +1582,6 @@ def create_receiving(function_table_data, function_tables_defs, exported_impleme
         else:
           receiving += '\n'.join(make_export_wrappers(module_exports, delay_assignment))
     else:
-      receiving += 'Module["asm"] = asm;\n'
       receiving += '\n'.join(make_export_wrappers(module_exports, delay_assignment))
 
   if shared.Settings.EXPORT_FUNCTION_TABLES and not shared.Settings.WASM:
@@ -2458,7 +2457,6 @@ def create_receiving_wasm(exports, initializers):
       else:
         receiving += make_export_wrappers(exports, delay_assignment)
   else:
-    receiving.append('Module["asm"] = asm;')
     receiving += make_export_wrappers(exports, delay_assignment)
 
   return '\n'.join(receiving) + '\n'