+++ src/postamble.js
@@ -6,7 +6,10 @@
 
 // === Auto-generated postamble setup entry stuff ===
 
+#if !WASM_BACKEND && !WASM
+// asm.js startup is synchronous
 Module['asm'] = asm;
+#endif
 
 {{{ exportRuntime() }}}
 
@@ -547,4 +550,8 @@ var workerResponded = false, workerCallbackId = -1;
   }
 })();
 
+#endif
+
+#if STANDALONE_WASM && ASSERTIONS && !WASM_BIGINT
+err('warning: running JS from STANDALONE_WASM without WASM_BIGINT will fail if a syscall with i64 is used (in standalone mode we cannot legalize syscalls)');
 #endif