+++ tests/test_sanity.py
@@ -22,7 +22,7 @@ from tools.shared import expected_llvm_version, Cache, Settings
 from tools import jsrun, shared, system_libs
 
 SANITY_FILE = CONFIG_FILE + '_sanity'
-commands = [[PYTHON, EMCC], [PYTHON, path_from_root('tests', 'runner.py'), 'blahblah']]
+commands = [[EMCC], [PYTHON, path_from_root('tests', 'runner.py'), 'blahblah']]
 
 
 def restore():
@@ -118,18 +118,19 @@ class sanity(RunnerCore):
     # Unlike the other test suites we explicitly don't want to be skipping
     # the sanity checks here
     del os.environ['EMCC_SKIP_SANITY_CHECK']
+
+    assert os.path.exists(CONFIG_FILE), 'To run these tests, we need a (working!) %s file to already exist' % EM_CONFIG
     shutil.copyfile(CONFIG_FILE, CONFIG_FILE + '_backup')
 
     print()
     print('Running sanity checks.')
-    print('WARNING: This will modify %s, and in theory can break it although it should be restored properly. A backup will be saved in %s_backup' % (EM_CONFIG, EM_CONFIG))
+    print('WARNING: This will modify %s, and in theory can break it although it should be restored properly. A backup will be saved in %s_backup' % (CONFIG_FILE, CONFIG_FILE))
     print()
     print('>>> the original settings file is:')
-    print(open(os.path.expanduser('~/.emscripten')).read())
+    print(open(CONFIG_FILE).read())
     print('<<<')
     print()
 
-    assert os.path.exists(CONFIG_FILE), 'To run these tests, we need a (working!) %s file to already exist' % EM_CONFIG
     assert 'EMCC_DEBUG' not in os.environ, 'do not run sanity checks in debug mode!'
     assert 'EMCC_WASM_BACKEND' not in os.environ, 'do not force wasm backend either way in sanity checks!'
 
@@ -147,14 +148,12 @@ class sanity(RunnerCore):
     super(sanity, self).tearDown()
     print('time:', time.time() - self.start_time)
 
-  def do(self, command):
+  def do(self, command, env=None):
     print('Running: ' + ' '.join(command))
     if type(command) is not list:
       command = [command]
-    if command[0] == EMCC:
-      command = [PYTHON] + command
 
-    return run_process(command, stdout=PIPE, stderr=STDOUT, check=False).stdout
+    return run_process(command, stdout=PIPE, stderr=STDOUT, check=False, env=env).stdout
 
   def check_working(self, command, expected=None):
     if type(command) is not list:
@@ -184,21 +183,24 @@ class sanity(RunnerCore):
         with open(os.path.join(temp_bin, name), 'w') as f:
           os.fchmod(f.fileno(), stat.S_IRWXU)
 
+      env = os.environ.copy()
+      if 'EM_CONFIG' in env:
+        del env['EM_CONFIG']
+
       try:
         temp_bin = tempfile.mkdtemp()
-        old_environ_path = os.environ['PATH']
-        os.environ['PATH'] = temp_bin + os.pathsep + old_environ_path
         make_executable('llvm-dis')
         make_executable('node')
         make_executable('python2')
-        output = self.do(command)
+        env['PATH'] = temp_bin + os.pathsep + os.environ['PATH']
+        output = self.do(command, env=env)
       finally:
-        os.environ['PATH'] = old_environ_path
         shutil.rmtree(temp_bin)
 
+      default_config = shared.embedded_config
       self.assertContained('Welcome to Emscripten!', output)
       self.assertContained('This is the first time any of the Emscripten tools has been run.', output)
-      self.assertContained('A settings file has been copied to %s, at absolute path: %s' % (EM_CONFIG, CONFIG_FILE), output)
+      self.assertContained('A settings file has been copied to %s, at absolute path: %s' % (default_config, default_config), output)
       self.assertContained('It contains our best guesses for the important paths, which are:', output)
       self.assertContained('LLVM_ROOT', output)
       self.assertContained('NODE_JS', output)
@@ -212,10 +214,8 @@ class sanity(RunnerCore):
       self.assertContained('Please edit the file if any of those are incorrect', output)
       self.assertContained('This command will now exit. When you are done editing those paths, re-run it.', output)
       assert output.split()[-1].endswith('===='), 'We should have stopped: ' + output
-      config_file = open(CONFIG_FILE).read()
+      config_file = open(default_config).read()
       template_file = open(path_from_root('tools', 'settings_template.py')).read()
-      self.assertNotContained('~/.emscripten', config_file)
-      self.assertContained('~/.emscripten', template_file)
       self.assertNotContained('{{{', config_file)
       self.assertNotContained('}}}', config_file)
       self.assertContained('{{{', template_file)
@@ -226,20 +226,21 @@ class sanity(RunnerCore):
       # The guessed config should be ok
       # XXX This depends on your local system! it is possible `which` guesses wrong
       # try_delete('a.out.js')
-      # output = run_process([PYTHON, EMCC, path_from_root('tests', 'hello_world.c')], stdout=PIPE, stderr=PIPE).output
+      # output = run_process([EMCC, path_from_root('tests', 'hello_world.c')], stdout=PIPE, stderr=PIPE).output
       # self.assertContained('hello, world!', run_js('a.out.js'), output)
 
       # Second run, with bad EM_CONFIG
       for settings in ['blah', 'LLVM_ROOT="blarg"; JS_ENGINES=[]; NODE_JS=[]; SPIDERMONKEY_ENGINE=[]']:
-        f = open(CONFIG_FILE, 'w')
+        f = open(default_config, 'w')
         f.write(settings)
         f.close()
-        output = self.do(command)
+        output = self.do(command, env=env)
 
         if 'LLVM_ROOT' not in settings:
-          self.assertContained('Error in evaluating %s' % EM_CONFIG, output)
+          self.assertContained('Error in evaluating %s' % default_config, output)
         elif 'runner.py' not in ' '.join(command):
           self.assertContained('error:', output) # sanity check should fail
+      try_delete(default_config)
 
   def test_llvm(self):
     LLVM_WARNING = 'LLVM version appears incorrect'
@@ -425,7 +426,7 @@ fi
 
     wipe()
     with env_modify({'EM_CONFIG': config}):
-      run_process([PYTHON, EMCC, 'main.cpp', '-o', 'a.out.js'])
+      run_process([EMCC, 'main.cpp', '-o', 'a.out.js'])
 
     self.assertContained('hello from emcc with no config file', run_js('a.out.js'))
 
@@ -438,7 +439,7 @@ fi
       self.assertEqual(os.listdir(Cache.dirname), [])
 
   def ensure_cache(self):
-    self.do([PYTHON, EMCC, '-O2', path_from_root('tests', 'hello_world.c')])
+    self.do([EMCC, '-O2', path_from_root('tests', 'hello_world.c')])
 
   def test_emcc_caching(self):
     BUILDING_MESSAGE = 'generating system library: X'
@@ -467,7 +468,7 @@ fi
     self.ensure_cache()
     self.assertTrue(os.path.exists(Cache.dirname))
     self.assertTrue(os.path.exists(Cache.root_dirname))
-    output = self.do([PYTHON, EMCC, '--clear-cache'])
+    output = self.do([EMCC, '--clear-cache'])
     self.assertIn(ERASING_MESSAGE, output)
     self.assertIn(SANITY_MESSAGE, output)
     self.assertCacheEmpty()
@@ -478,7 +479,7 @@ fi
     make_fake_llc(self.in_dir('fake', 'bin', 'llc'), 'js - JavaScript (asm.js, emscripten)')
     with env_modify({'EM_LLVM_ROOT': self.in_dir('fake', 'bin')}):
       self.assertTrue(os.path.exists(Cache.dirname))
-      output = self.do([PYTHON, EMCC])
+      output = self.do([EMCC])
       self.assertIn(ERASING_MESSAGE, output)
       self.assertCacheEmpty()
 
@@ -491,11 +492,11 @@ fi
     self.assertTrue(os.path.exists(Cache.root_dirname))
     # changing config file should not clear cache
     add_to_config('FROZEN_CACHE = True')
-    self.do([PYTHON, EMCC])
+    self.do([EMCC])
     self.assertTrue(os.path.exists(Cache.dirname))
     self.assertTrue(os.path.exists(Cache.root_dirname))
     # building libraries is disallowed
-    output = self.do([PYTHON, EMBUILDER, 'build', 'libemmalloc'])
+    output = self.do([EMBUILDER, 'build', 'libemmalloc'])
     self.assertIn('FROZEN_CACHE disallows building system libs', output)
 
   # Test that if multiple processes attempt to access or build stuff to the
@@ -516,7 +517,7 @@ fi
       tasks = []
       num_times_libc_was_built = 0
       for i in range(3):
-        p = run_process([PYTHON, EMCC, 'test.c', '-o', '%d.js' % i], stderr=STDOUT, stdout=PIPE)
+        p = run_process([EMCC, 'test.c', '-o', '%d.js' % i], stderr=STDOUT, stdout=PIPE)
         tasks += [p]
       for p in tasks:
         print('stdout:\n', p.stdout)
@@ -550,7 +551,7 @@ fi
     temp_dir = tempfile.mkdtemp(prefix='emscripten_temp_')
 
     with chdir(temp_dir):
-      run_process([PYTHON, EMCC, '--em-config', custom_config_filename] + MINIMAL_HELLO_WORLD + ['-O2'])
+      run_process([EMCC, '--em-config', custom_config_filename] + MINIMAL_HELLO_WORLD + ['-O2'])
       result = run_js('a.out.js')
 
     self.assertContained('hello, world!', result)
@@ -569,7 +570,7 @@ fi
 
     # listing ports
 
-    out = self.do([PYTHON, EMCC, '--show-ports'])
+    out = self.do([EMCC, '--show-ports'])
     assert 'Available ports:' in out, out
     assert 'SDL2' in out, out
     assert 'SDL2_image' in out, out
@@ -582,12 +583,12 @@ fi
     PORTS_DIR = system_libs.Ports.get_dir()
 
     for i in [0, 1]:
-      self.do([PYTHON, EMCC, '--clear-cache'])
+      self.do([EMCC, '--clear-cache'])
       print(i)
       if i == 0:
         try_delete(PORTS_DIR)
       else:
-        self.do([PYTHON, EMCC, '--clear-ports'])
+        self.do([EMCC, '--clear-ports'])
       assert not os.path.exists(PORTS_DIR)
 
       # Building a file that doesn't need ports should not trigger anything
@@ -814,11 +815,11 @@ fi
 
   def test_embuilder_force(self):
     restore_and_set_up()
-    self.do([PYTHON, EMBUILDER, 'build', 'libemmalloc'])
+    self.do([EMBUILDER, 'build', 'libemmalloc'])
     # Second time it should not generate anything
-    self.assertNotContained('generating system library', self.do([PYTHON, EMBUILDER, 'build', 'libemmalloc']))
+    self.assertNotContained('generating system library', self.do([EMBUILDER, 'build', 'libemmalloc']))
     # Unless --force is specified
-    self.assertContained('generating system library', self.do([PYTHON, EMBUILDER, 'build', 'libemmalloc', '--force']))
+    self.assertContained('generating system library', self.do([EMBUILDER, 'build', 'libemmalloc', '--force']))
 
   def test_embuilder_wasm_backend(self):
     if not Settings.WASM_BACKEND:
@@ -826,11 +827,11 @@ fi
     restore_and_set_up()
     root_cache = os.path.expanduser('~/.emscripten_cache')
     # the --lto flag makes us build wasm-bc
-    self.do([PYTHON, EMCC, '--clear-cache'])
-    run_process([PYTHON, EMBUILDER, 'build', 'libemmalloc'])
+    self.do([EMCC, '--clear-cache'])
+    run_process([EMBUILDER, 'build', 'libemmalloc'])
     self.assertExists(os.path.join(root_cache, 'wasm-obj'))
-    self.do([PYTHON, EMCC, '--clear-cache'])
-    run_process([PYTHON, EMBUILDER, 'build', 'libemmalloc', '--lto'])
+    self.do([EMCC, '--clear-cache'])
+    run_process([EMBUILDER, 'build', 'libemmalloc', '--lto'])
     self.assertExists(os.path.join(root_cache, 'wasm-bc'))
 
   def test_binaryen_version(self):